+++ packages/form-data-parser/examples/node/server.js
@@ -3,13 +3,13 @@ import * as http from 'node:http';
 import * as os from 'node:os';
 import * as path from 'node:path';
 
-import { LocalFileStorage } from '@mjackson/file-storage/local';
+import { LocalFileStorage } from '@remix-run/file-storage/local';
 import {
   MultipartParseError,
   MaxFileSizeExceededError,
   parseFormData,
-} from '@mjackson/form-data-parser';
-import { createRequestListener } from '@mjackson/node-fetch-server';
+} from '@remix-run/form-data-parser';
+import { createRequestListener } from '@remix-run/node-fetch-server';
 
 const PORT = 3000;