+++ packages/form-data-parser/src/form-data-parser.ts
@@ -11,4 +11,4 @@ export {
   MultipartParseError,
   MaxHeaderSizeExceededError,
   MaxFileSizeExceededError,
-} from '@mjackson/multipart-parser';
+} from '@remix-run/multipart-parser';