+++ packages/tar-parser/src/lib/tar.ts
@@ -159,7 +159,7 @@ export type ParseTarOptions = ParseTarHeaderOptions;
  * Parse a tar archive and call the given handler for each entry it contains.
  *
  * ```ts
- * import { parseTar } from '@mjackson/tar-parser';
+ * import { parseTar } from '@remix-run/tar-parser';
  *
  * await parseTar(archive, (entry) => {
  *  console.log(entry.name);