+++ scripts/utils/fs.js
@@ -19,6 +19,13 @@ export function readFile(filename, encoding = 'utf-8') {
   }
 }
 
+/** @type (error: unknown) => error is { code: string } */
+function isFsError(error) {
+  return (
+    typeof error === 'object' && error != null && 'code' in error && typeof error.code === 'string'
+  );
+}
+
 /** @type (filename: string, data: string) => void */
 export function writeFile(filename, data) {
   fs.writeFileSync(filename, data);
@@ -32,11 +39,4 @@ export function readJson(filename) {
 /** @type (filename: string, data: any) => void */
 export function writeJson(filename, data) {
   writeFile(filename, JSON.stringify(data, null, 2) + '\n');
-}
-
-/** @type (error: unknown) => error is { code: string } */
-function isFsError(error) {
-  return (
-    typeof error === 'object' && error != null && 'code' in error && typeof error.code === 'string'
-  );
 }