+++ packages/route-pattern/CHANGELOG.md
@@ -1,5 +1,7 @@
 # `route-pattern` CHANGELOG
 
-This is the changelog for [`route-pattern`](https://github.com/mjackson/remix-the-web/tree/main/packages/route-pattern). It follows [semantic versioning](https://semver.org/).
+This is the changelog for [`route-pattern`](https://github.com/remix-run/remix/tree/v3/packages/route-pattern). It follows [semantic versioning](https://semver.org/).
 
 ## HEAD
+
+- Renamed package from `@mjackson/route-pattern` to `@remix-run/route-pattern`